move refresher to tabs component
