+++ src/app/pages/home/home.page.html
@@ -5,16 +5,6 @@
     </ion-toolbar>
   </ion-header>
 
-  <ion-refresher
-    slot="fixed"
-    [pullFactor]="0.5"
-    [pullMin]="100"
-    [pullMax]="200"
-    (ionRefresh)="handleRefresh($event)"
-  >
-    <ion-refresher-content> </ion-refresher-content>
-  </ion-refresher>
-
   <ion-item id="loginItem" *ngIf="!loggedIn" color="warning">
     <ion-label>You're not logged in</ion-label>
     <ion-button slot="end" fill="clear" (click)="handleLogin()"
