+++ src/app/pages/tabs/tabs.page.html
@@ -5,6 +5,16 @@
 </ion-header>
 
 <ion-content>
+  <ion-refresher
+    slot="fixed"
+    [pullFactor]="0.5"
+    [pullMin]="100"
+    [pullMax]="200"
+    (ionRefresh)="handleRefresh($event)"
+  >
+    <ion-refresher-content> </ion-refresher-content>
+  </ion-refresher>
+
   <ion-router-outlet></ion-router-outlet>
 
   <ion-tabs>
