+++ src/app/pages/tabs/tabs.page.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { JobsService } from 'src/app/services/Jobs/jobs.service';
+import { RepositoryService } from 'src/app/services/Repositories/repository.service';
 
 @Component({
   selector: 'app-tabs',
@@ -6,10 +8,30 @@ import { Component, OnInit } from '@angular/core';
   styleUrls: ['./tabs.page.scss'],
 })
 export class TabsPage implements OnInit {
+  constructor(
+    private jobsSrv: JobsService,
+    private repoSrv: RepositoryService
+  ) {}
 
-  constructor() { }
+  ngOnInit() {}
 
-  ngOnInit() {
+  async fetchMails() {
+    return new Promise<void>((resolve, reject) => {
+      this.jobsSrv.loadMails().subscribe({
+        next: (transaction) => {
+          console.log(transaction);
+        },
+        complete: () => {
+          this.repoSrv.save();
+          resolve();
+        },
+      });
+    });
   }
 
+  handleRefresh(event: Event) {
+    this.fetchMails().then(() => {
+      (event.target as HTMLIonRefresherElement).complete();
+    });
+  }
 }
